requirements.txt
Django==1.6
Markdown==2.4.1
Pillow==2.5.3
South==1.0
admin-tools-zinnia==0.1
argparse==1.2.1
args==0.1.0
beautifulsoup4==4.3.2
clint==0.3.7
configparser==3.3.0r2
django-admin-tools==0.5.2
django-app-namespace-template-loader==0.2
django-appconf==0.6
django-blog-zinnia==0.14.2
django-ckeditor==4.4.5
django-compressor==1.3
django-configurations==0.8
django-contrib-comments==1.5
django-extensions==1.3.3
django-filter==0.7
-e git+https://github.com/APSL/django-mailer-2.git@c330314eff85f690ec0630421b9cc9d7ba51947e#egg=django_mailer_2-master
-e git+https://github.com/APSL/django-mailviews.git@df4585e49bde0d3b01b6ae1cacdfe608a659bab5#egg=django_mailviews-master
django-modeltranslation==0.8b2
django-mptt==0.6.1
django-redis==3.4.0
django-tagging==0.3.2
django-tinymce==1.5.2
django-xmlrpc==0.1.5
djangorestframework==2.4.2
ipython==2.2.0
nose==1.3.4
psycopg2==2.5.2
pyparsing==2.0.2
pytz==2014.7
pyzmail==1.0.3
raven==5.0.0
redis==2.10.3
six==1.7.3
suds==0.4
wsgiref==0.1.2
zinnia-theme-html5==0.3.2
-e git+https://github.com/Fantomas42/zinnia-theme-bootstrap.git@2f25796c65692ac63a85a6df6227af2c209d2f58#egg=zinnia_theme_bootstrap-origin
